+++ imports/core/api/player-stages/augment.js
@@ -1,40 +1,72 @@
 import { PlayerRounds } from "../player-rounds/player-rounds";
 import { PlayerStages } from "./player-stages";
-import { Players } from "../players/players.js";
 import { submitPlayerStage, updatePlayerStageData } from "./methods";
-import { updatePlayerData } from "../players/methods";
+import { updateGameData } from "../games/methods.js";
+import { updatePlayerData } from "../players/methods.js";
 import { updatePlayerRoundData } from "../player-rounds/methods";
+import { updateRoundData } from "../rounds/methods.js";
+import { updateStageData } from "../stages/methods.js";
 
-const playerSet = playerId => (key, value) => {
+const gameSet = (gameId, append = false) => (key, value) => {
+  updateGameData.call({
+    gameId,
+    key,
+    value: JSON.stringify(value),
+    append,
+    noCallback: Meteor.isServer
+  });
+};
+const playerSet = (playerId, append = false) => (key, value) => {
   updatePlayerData.call({
     playerId,
     key,
-    value: JSON.stringify(value)
+    value: JSON.stringify(value),
+    append,
+    noCallback: Meteor.isServer
   });
 };
-const stageSet = playerStageId => (key, value) => {
+const stageSet = (playerStageId, append = false) => (key, value) => {
   updatePlayerStageData.call({
     playerStageId,
     key,
-    value: JSON.stringify(value)
+    value: JSON.stringify(value),
+    append,
+    noCallback: Meteor.isServer
   });
 };
 const stageSubmit = playerStageId => cb => {
-  submitPlayerStage.call({ playerStageId }, cb);
+  submitPlayerStage.call(
+    {
+      playerStageId,
+      noCallback: Meteor.isServer
+    },
+    cb
+  );
 };
-const roundSet = playerRoundId => (key, value) => {
+const roundSet = (playerRoundId, append = false) => (key, value) => {
   updatePlayerRoundData.call({
     playerRoundId,
     key,
-    value: JSON.stringify(value)
+    value: JSON.stringify(value),
+    append,
+    noCallback: Meteor.isServer
   });
 };
 
 // Once the operation has succeeded (no throw), set the value
-// "|| null" because undefined is unsupported.
+// undefined is not supported, null is, replace undefineds by nulls.
 const set = (obj, func) => (k, v) => {
-  func(k, v || null);
-  obj[k] = v || null;
+  const val = v === undefined ? null : v;
+  func(k, val);
+  obj[k] = val;
+};
+const append = (obj, func) => (k, v) => {
+  const val = v === undefined ? null : v;
+  func(k, val);
+  if (!obj[k]) {
+    obj[k] = [];
+  }
+  obj[k].push(val);
 };
 
 const nullFunc = () => {
@@ -46,11 +78,13 @@ export const augmentPlayerStageRound = (player, stage, round) => {
 
   player.get = key => player.data[key];
   player.set = set(player.data, playerSet(playerId));
+  player.append = append(player.data, playerSet(playerId, true));
 
   if (stage) {
     const playerStage = PlayerStages.findOne({ stageId: stage._id, playerId });
     stage.get = key => playerStage.data[key];
     stage.set = set(playerStage.data, stageSet(playerStage._id));
+    stage.append = append(playerStage.data, stageSet(playerStage._id, true));
     stage.submit = stageSubmit(playerStage._id, err => {
       if (!err) {
         stage.submitted = true;
@@ -63,16 +97,19 @@ export const augmentPlayerStageRound = (player, stage, round) => {
     const playerRound = PlayerRounds.findOne({ roundId: round._id, playerId });
     round.get = key => playerRound.data[key];
     round.set = set(playerRound.data, roundSet(playerRound._id));
+    round.append = append(playerRound.data, roundSet(playerRound._id, true));
   }
 };
 
 export const stubPlayerStageRound = (player, stage, round) => {
   player.get = nullFunc;
   player.set = nullFunc;
+  player.append = nullFunc;
 
   if (stage) {
     stage.get = nullFunc;
     stage.set = nullFunc;
+    stage.append = nullFunc;
     stage.submit = nullFunc;
     stage.submitted = false;
   }
@@ -80,17 +117,39 @@ export const stubPlayerStageRound = (player, stage, round) => {
   if (round) {
     round.get = nullFunc;
     round.set = nullFunc;
+    round.append = nullFunc;
   }
 };
 
-export const augmentStageRound = (stage, round) => {
+export const augmentGameStageRound = (game, stage, round) => {
+  if (game) {
+    game.get = key => game.data[key];
+    game.set = set(game.data, gameSet(game._id));
+    game.append = append(game.data, gameSet(game._id, true));
+  }
+
   if (stage) {
     stage.get = key => {
-      return state.data[key];
+      return stage.data[key];
     };
-    stage.set = (key, value) => {
-      throw "You cannot update stage data at the moment";
-    };
+    stage.set = set(stage.data, (key, value) => {
+      updateStageData.call({
+        stageId: stage._id,
+        key,
+        value: JSON.stringify(value),
+        append: false,
+        noCallback: Meteor.isServer
+      });
+    });
+    stage.append = append(stage.data, (key, value) => {
+      updateStageData.call({
+        stageId: stage._id,
+        key,
+        value: JSON.stringify(value),
+        append: true,
+        noCallback: Meteor.isServer
+      });
+    });
     stage.submit = () => {
       throw "You cannot submit the entire stage at the moment";
     };
@@ -100,16 +159,33 @@ export const augmentStageRound = (stage, round) => {
     round.get = key => {
       return round.data[key];
     };
-    round.set = (key, value) => {
-      throw "You cannot update round data at the moment";
-    };
+    round.set = set(round.data, (key, value) => {
+      updateRoundData.call({
+        roundId: round._id,
+        key,
+        value: JSON.stringify(value),
+        append: false,
+        noCallback: Meteor.isServer
+      });
+    });
+    round.append = append(round.data, (key, value) => {
+      updateRoundData.call({
+        roundId: round._id,
+        key,
+        value: JSON.stringify(value),
+        append: true,
+        noCallback: Meteor.isServer
+      });
+    });
   }
 };
 
 export const stubStageRound = (stage, round) => {
   stage.get = nullFunc;
   stage.set = nullFunc;
+  stage.append = nullFunc;
   stage.submit = nullFunc;
   round.get = nullFunc;
   round.set = nullFunc;
+  round.append = nullFunc;
 };