+++ imports/core/api/stages/methods.js
@@ -1,17 +1,20 @@
 import { ValidatedMethod } from "meteor/mdg:validated-method";
 import SimpleSchema from "simpl-schema";
-import { PlayerRounds } from "./player-rounds";
 
-let callOnChange;
+import { Stages } from "./stages.js";
+
+let callOnChange, playerIdForConn;
 if (Meteor.isServer) {
+  playerIdForConn = require("../../startup/server/connections.js")
+    .playerIdForConn;
   callOnChange = require("../server/onchange").callOnChange;
 }
 
-export const updatePlayerRoundData = new ValidatedMethod({
-  name: "PlayerRounds.methods.updateData",
+export const updateStageData = new ValidatedMethod({
+  name: "Stages.methods.updateData",
 
   validate: new SimpleSchema({
-    playerRoundId: {
+    stageId: {
       type: String,
       regEx: SimpleSchema.RegEx.Id
     },
@@ -31,27 +34,27 @@ export const updatePlayerRoundData = new ValidatedMethod({
     }
   }).validator(),
 
-  run({ playerRoundId, key, value, append, noCallback }) {
-    const playerRound = PlayerRounds.findOne(playerRoundId);
-    if (!playerRound) {
-      throw new Error("playerRound not found");
+  run({ stageId, key, value, append, noCallback }) {
+    const stage = Stages.findOne(stageId);
+    if (!stage) {
+      throw new Error("stage not found");
     }
-    // TODO check can update this record playerRound
+    // TODO check can update this record stage
 
     const val = JSON.parse(value);
     let update = { [`data.${key}`]: val };
     const modifier = append ? { $push: update } : { $set: update };
 
-    PlayerRounds.update(playerRoundId, modifier, { autoConvert: false });
+    Stages.update(stageId, modifier, { autoConvert: false });
 
     if (Meteor.isServer && !noCallback) {
       callOnChange({
-        playerId: playerRound.playerId,
-        playerRoundId,
-        playerRound,
+        playerId: playerIdForConn(this.connection),
+        stageId,
+        stage,
         key,
         value: val,
-        prevValue: playerRound.data && playerRound.data[key],
+        prevValue: stage.data && stage.data[key],
         append
       });
     }